Config: Add templated configGet() overloads.
This is for a way better "Just do what I mean"-expierence, while still
holding the "WTF dosen't it just do it?" a bit farther off...

configGet( "Foo", 5 )         => returns int
configGet( "Foo", "Bar" )     => returns QString
configGet( "Foo", QString() ) => returns QString
configGetV( "Foo", "Bar" )    => returns QVariant (toString() needed)
configGetV( "Foo", 5 )        => returns QVariant (toInt() needed)

+++ src/Modules/History/HistoryView.cpp
@@ -165,7 +165,7 @@ void HistoryView::initSplitters()
 	mDetails->hide();
 	mDetails->setParent( NULL );
 
-	int i = configGet( "SplitLayout", 1 ).toInt();
+	int i = configGet( "SplitLayout", 1 );
 	switch( i )
 	{
 	default:
